requirements.txt
asgiref @ file:///tmp/build/80754af9/asgiref_1594338739818/work
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django @ file:///tmp/build/80754af9/django_1600355803933/work
django-cors-headers==3.5.0
django-filter==2.4.0
django-heroku==0.3.1
django-rest-swagger==2.2.0
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
openapi-codec==1.3.2
PyJWT==1.7.1
pytz==2020.1
requests==2.24.0
simplejson==3.17.2
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.10
whitenoise==5.2.0
